I fixed the error
Before, the server would not connect, the IP and PORT were invalid. After a bid of work, I made it so that the user put in their own IP (assuming they know how). If the user does not know their IPv4, the instructions will be in the Github README.md.